+++ src/test/java/org/alliancegenome/curation_api/bulkupload/fms/AGMBulkUploadFmsITCase.java
@@ -1,9 +1,7 @@
-package org.alliancegenome.curation_api.bulkupload.ontology.fms;
+package org.alliancegenome.curation_api.bulkupload.fms;
 
-import java.io.IOException;
 import java.nio.file.*;
 
-import org.alliancegenome.curation_api.model.entities.Gene;
 import org.alliancegenome.curation_api.resources.TestElasticSearchResource;
 import org.junit.jupiter.api.*;
 
@@ -19,8 +17,8 @@ import static org.hamcrest.Matchers.is;
 @QuarkusTestResource(TestElasticSearchResource.Initializer.class)
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-@Order(2)
-public class AlleleBulkUploadFmsITCase {
+@Order(3)
+public class AGMBulkUploadFmsITCase {
     
     @BeforeEach
     public void init() {
@@ -32,15 +30,15 @@ public class AlleleBulkUploadFmsITCase {
 
     @Test
     @Order(1)
-    public void alleleBulkUploadCheckFields() throws Exception {
-        String content = Files.readString(Path.of("src/test/resources/bulk/fms/02_allele/01_all_fields.json"));
+    public void agmBulkUploadCheckFields() throws Exception {
+        String content = Files.readString(Path.of("src/test/resources/bulk/fms/03_affected_genomic_model/01_all_fields.json"));
 
         // upload file
         RestAssured.given().
             contentType("application/json").
             body(content).
             when().
-            post("/api/allele/bulk/allelefile").
+            post("/api/agm/bulk/agmfile").
             then().
             statusCode(200);
 
@@ -50,31 +48,35 @@ public class AlleleBulkUploadFmsITCase {
             when().
             header("Content-Type", "application/json").
             body("{}").
-            post("/api/allele/find?limit=10&page=0").
+            post("/api/agm/find?limit=10&page=0").
             then().
             statusCode(200).
             body("totalResults", is(1)).
             body("results", hasSize(1)).
-            body("results[0].curie", is("TEST:TestAllele00001")).
+            body("results[0].curie", is("TEST:TestAGM00001")).
             body("results[0].taxon.curie", is("NCBITaxon:10090")).
-            body("results[0].symbol", is("Test<sup>allele</sup>")).
-            body("results[0].synonyms[0].name", is("test<sup>+</sup>")).
-            body("results[0].crossReferences[0].curie", is("TEST:TestAllele00001")).
-            body("results[0].secondaryIdentifiers[0]", is("TEST:A1")).
-            body("results[0].description", is("Test description of allele"));
+            body("results[0].name", is( "Test AGM 1")).
+            body("results[0].synonyms[0].name", is("TAGM1")).
+            body("results[0].crossReferences[0].curie", is("TEST:TestAGM00001")).
+            body("results[0].secondaryIdentifiers[0]", is("TEST:AGM1")).
+            // body("results[0].components[0].allele.curie", is(TESTALLELE)).
+            // body("results[0].components[0].zygosity", is("GENO:0000136")).
+            // body("results[0].parentalPopulations[0]", is("TEST:TestComponent00001")).
+            // body("results[0].sequenceTargetingReagents[0].curie", is("TEST:TestSTR00001")).
+            body("results[0].subtype", is("genotype"));
     }
     
     @Test
     @Order(2)
-    public void alleleBulkUploadNoAlleleObjectRelations() throws Exception {
-        String content = Files.readString(Path.of("src/test/resources/bulk/fms/02_allele/02_no_allele_object_relations.json"));
+    public void agmBulkUploadNoComponents() throws Exception {
+        String content = Files.readString(Path.of("src/test/resources/bulk/fms/03_affected_genomic_model/02_no_agm_components.json"));
         
         // upload file
         RestAssured.given().
             contentType("application/json").
             body(content).
             when().
-            post("/api/allele/bulk/allelefile").
+            post("/api/agm/bulk/agmfile").
             then().
             statusCode(200);
         
@@ -83,7 +85,7 @@ public class AlleleBulkUploadFmsITCase {
             when().
             header("Content-Type", "application/json").
             body("{}").
-            post("/api/allele/find?limit=10&page=0").
+            post("/api/agm/find?limit=10&page=0").
             then().
             statusCode(200).
             body("totalResults", is(2));
@@ -91,15 +93,15 @@ public class AlleleBulkUploadFmsITCase {
 
     @Test
     @Order(3)
-    public void alleleBulkUploadNoGeneAlleleObjectRelation() throws Exception {
-        String content = Files.readString(Path.of("src/test/resources/bulk/fms/02_allele/03_no_gene_allele_object_relation.json"));
+    public void agmBulkUploadNoComponentAlleleId() throws Exception {
+        String content = Files.readString(Path.of("src/test/resources/bulk/fms/03_affected_genomic_model/03_no_agm_components_allele_id.json"));
         
         // upload file
         RestAssured.given().
             contentType("application/json").
             body(content).
             when().
-            post("/api/allele/bulk/allelefile").
+            post("/api/agm/bulk/agmfile").
             then().
             statusCode(200);
         
@@ -108,23 +110,73 @@ public class AlleleBulkUploadFmsITCase {
             when().
             header("Content-Type", "application/json").
             body("{}").
-            post("/api/allele/find?limit=10&page=0").
+            post("/api/agm/find?limit=10&page=0").
+            then().
+            statusCode(200).
+            body("totalResults", is(2));
+    }
+
+    @Test
+    @Order(4)
+    public void agmBulkUploadNoComponentZygosity() throws Exception {
+        String content = Files.readString(Path.of("src/test/resources/bulk/fms/03_affected_genomic_model/04_no_agm_components_zygosity.json"));
+        
+        // upload file
+        RestAssured.given().
+            contentType("application/json").
+            body(content).
+            when().
+            post("/api/agm/bulk/agmfile").
+            then().
+            statusCode(200);
+        
+        // check entity count
+        RestAssured.given().
+            when().
+            header("Content-Type", "application/json").
+            body("{}").
+            post("/api/agm/find?limit=10&page=0").
+            then().
+            statusCode(200).
+            body("totalResults", is(2));
+    }
+
+    @Test
+    @Order(5)
+    public void agmBulkUploadNoCrossReference() throws Exception {
+        String content = Files.readString(Path.of("src/test/resources/bulk/fms/03_affected_genomic_model/05_no_cross_reference.json"));
+        
+        // upload file
+        RestAssured.given().
+            contentType("application/json").
+            body(content).
+            when().
+            post("/api/agm/bulk/agmfile").
+            then().
+            statusCode(200);
+        
+        // check entity count
+        RestAssured.given().
+            when().
+            header("Content-Type", "application/json").
+            body("{}").
+            post("/api/agm/find?limit=10&page=0").
             then().
             statusCode(200).
             body("totalResults", is(3));
     }
 
     @Test
-    @Order(4)
-    public void alleleBulkUploadNoConstructAlleleObjectRelation() throws Exception {
-        String content = Files.readString(Path.of("src/test/resources/bulk/fms/02_allele/04_no_construct_allele_object_relation.json"));
+    @Order(6)
+    public void agmBulkUploadNoCrossReferenceId() throws Exception {
+        String content = Files.readString(Path.of("src/test/resources/bulk/fms/03_affected_genomic_model/06_no_cross_reference_id.json"));
         
         // upload file
         RestAssured.given().
             contentType("application/json").
             body(content).
             when().
-            post("/api/allele/bulk/allelefile").
+            post("/api/agm/bulk/agmfile").
             then().
             statusCode(200);
         
@@ -133,23 +185,48 @@ public class AlleleBulkUploadFmsITCase {
             when().
             header("Content-Type", "application/json").
             body("{}").
-            post("/api/allele/find?limit=10&page=0").
+            post("/api/agm/find?limit=10&page=0").
+            then().
+            statusCode(200).
+            body("totalResults", is(3)); //no entity added due to missing ID
+    }
+
+    @Test
+    @Order(7)
+    public void agmBulkUploadNoCrossReferencePages() throws Exception {
+        String content = Files.readString(Path.of("src/test/resources/bulk/fms/03_affected_genomic_model/07_no_cross_reference_pages.json"));
+        
+        // upload file
+        RestAssured.given().
+            contentType("application/json").
+            body(content).
+            when().
+            post("/api/agm/bulk/agmfile").
+            then().
+            statusCode(200);
+        
+        // check entity count
+        RestAssured.given().
+            when().
+            header("Content-Type", "application/json").
+            body("{}").
+            post("/api/agm/find?limit=10&page=0").
             then().
             statusCode(200).
             body("totalResults", is(4));
     }
 
     @Test
-    @Order(5)
-    public void alleleBulkUploadNoAlleleObjectRelationAssociationType() throws Exception {
-        String content = Files.readString(Path.of("src/test/resources/bulk/fms/02_allele/05_no_allele_object_relation_association_type.json"));
+    @Order(8)
+    public void agmBulkUploadNoName() throws Exception {
+        String content = Files.readString(Path.of("src/test/resources/bulk/fms/03_affected_genomic_model/08_no_name.json"));
         
         // upload file
         RestAssured.given().
             contentType("application/json").
             body(content).
             when().
-            post("/api/allele/bulk/allelefile").
+            post("/api/agm/bulk/agmfile").
             then().
             statusCode(200);
         
@@ -158,23 +235,48 @@ public class AlleleBulkUploadFmsITCase {
             when().
             header("Content-Type", "application/json").
             body("{}").
-            post("/api/allele/find?limit=10&page=0").
+            post("/api/agm/find?limit=10&page=0").
+            then().
+            statusCode(200).
+            body("totalResults", is(4)); // Name is required field so entity skipped in load
+    }
+
+    @Test
+    @Order(9)
+    public void agmBulkUploadNoParentalPopulationIds() throws Exception {
+        String content = Files.readString(Path.of("src/test/resources/bulk/fms/03_affected_genomic_model/09_no_parental_population_ids.json"));
+        
+        // upload file
+        RestAssured.given().
+            contentType("application/json").
+            body(content).
+            when().
+            post("/api/agm/bulk/agmfile").
+            then().
+            statusCode(200);
+        
+        // check entity count
+        RestAssured.given().
+            when().
+            header("Content-Type", "application/json").
+            body("{}").
+            post("/api/agm/find?limit=10&page=0").
             then().
             statusCode(200).
             body("totalResults", is(5));
     }
 
     @Test
-    @Order(6)
-    public void alleleBulkUploadNoAlleleObjectRelationEntity() throws Exception {
-        String content = Files.readString(Path.of("src/test/resources/bulk/fms/02_allele/06_no_allele_object_relation_entity.json"));
+    @Order(10)
+    public void agmBulkUploadNoPrimaryId() throws Exception {
+        String content = Files.readString(Path.of("src/test/resources/bulk/fms/03_affected_genomic_model/10_no_primary_id.json"));
         
         // upload file
         RestAssured.given().
             contentType("application/json").
             body(content).
             when().
-            post("/api/allele/bulk/allelefile").
+            post("/api/agm/bulk/agmfile").
             then().
             statusCode(200);
         
@@ -183,24 +285,23 @@ public class AlleleBulkUploadFmsITCase {
             when().
             header("Content-Type", "application/json").
             body("{}").
-            post("/api/allele/find?limit=10&page=0").
+            post("/api/agm/find?limit=10&page=0").
             then().
             statusCode(200).
-            body("totalResults", is(5));
+            body("totalResults", is(5)); // PrimaryID is required field so entity skipped in load
     }
 
-    // TODO: adjust count (and subsequent test counts) once validation of xref pages in place
     @Test
-    @Order(7)
-    public void alleleBulkUploadNoCrossReferences() throws Exception {
-        String content = Files.readString(Path.of("src/test/resources/bulk/fms/02_allele/07_no_cross_references.json"));
+    @Order(11)
+    public void agmBulkUploadNoSecondaryIds() throws Exception {
+        String content = Files.readString(Path.of("src/test/resources/bulk/fms/03_affected_genomic_model/11_no_secondary_ids.json"));
         
         // upload file
         RestAssured.given().
             contentType("application/json").
             body(content).
             when().
-            post("/api/allele/bulk/allelefile").
+            post("/api/agm/bulk/agmfile").
             then().
             statusCode(200);
         
@@ -209,48 +310,23 @@ public class AlleleBulkUploadFmsITCase {
             when().
             header("Content-Type", "application/json").
             body("{}").
-            post("/api/allele/find?limit=10&page=0").
+            post("/api/agm/find?limit=10&page=0").
             then().
             statusCode(200).
             body("totalResults", is(6));
     }
 
     @Test
-    @Order(8)
-    public void alleleBulkUploadNoCrossReferenceId() throws Exception {
-        String content = Files.readString(Path.of("src/test/resources/bulk/fms/02_allele/08_no_cross_reference_id.json"));
+    @Order(12)
+    public void agmBulkUploadNoSequenceTargetingReagentIds() throws Exception {
+        String content = Files.readString(Path.of("src/test/resources/bulk/fms/03_affected_genomic_model/12_no_str_ids.json"));
         
         // upload file
         RestAssured.given().
             contentType("application/json").
             body(content).
             when().
-            post("/api/allele/bulk/allelefile").
-            then().
-            statusCode(200);
-        
-        // check entity count
-            RestAssured.given().
-                when().
-                header("Content-Type", "application/json").
-                body("{}").
-                post("/api/allele/find?limit=10&page=0").
-                then().
-                statusCode(200).
-                body("totalResults", is(6)); // no entity added due to missing ID
-    }
-
-    @Test
-    @Order(9)
-    public void alleleBulkUploadNoCrossReferencePages() throws Exception {
-        String content = Files.readString(Path.of("src/test/resources/bulk/fms/02_allele/09_no_cross_reference_pages.json"));
-        
-        // upload file
-        RestAssured.given().
-            contentType("application/json").
-            body(content).
-            when().
-            post("/api/allele/bulk/allelefile").
+            post("/api/agm/bulk/agmfile").
             then().
             statusCode(200);
         
@@ -259,98 +335,23 @@ public class AlleleBulkUploadFmsITCase {
             when().
             header("Content-Type", "application/json").
             body("{}").
-            post("/api/allele/find?limit=10&page=0").
+            post("/api/agm/find?limit=10&page=0").
             then().
             statusCode(200).
             body("totalResults", is(7));
     }
 
     @Test
-    @Order(10)
-    public void alleleBulkUploadNoDescription() throws Exception {
-        String content = Files.readString(Path.of("src/test/resources/bulk/fms/02_allele/10_no_description.json"));
-        
-        // upload file
-        RestAssured.given().
-            contentType("application/json").
-            body(content).
-            when().
-            post("/api/allele/bulk/allelefile").
-            then().
-            statusCode(200);
-        
-        // check entity count
-        RestAssured.given().
-            when().
-            header("Content-Type", "application/json").
-            body("{}").
-            post("/api/allele/find?limit=10&page=0").
-            then().
-            statusCode(200).
-            body("totalResults", is(8)); // PrimaryID is required field so entity skipped in load
-    }
-
-    @Test
-    @Order(11)
-    public void alleleBulkUploadNoPrimaryId() throws Exception {
-        String content = Files.readString(Path.of("src/test/resources/bulk/fms/02_allele/11_no_primary_id.json"));
-        
-        // upload file
-        RestAssured.given().
-            contentType("application/json").
-            body(content).
-            when().
-            post("/api/allele/bulk/allelefile").
-            then().
-            statusCode(200);
-        
-        // check entity count
-        RestAssured.given().
-            when().
-            header("Content-Type", "application/json").
-            body("{}").
-            post("/api/allele/find?limit=10&page=0").
-            then().
-            statusCode(200).
-            body("totalResults", is(8));
-    }
-
-    @Test
-    @Order(12)
-    public void alleleBulkUploadNoSymbol() throws Exception {
-        String content = Files.readString(Path.of("src/test/resources/bulk/fms/02_allele/12_no_symbol.json"));
-        
-        // upload file
-        RestAssured.given().
-            contentType("application/json").
-            body(content).
-            when().
-            post("/api/allele/bulk/allelefile").
-            then().
-            statusCode(200);
-        
-        // check entity count
-        RestAssured.given().
-            when().
-            header("Content-Type", "application/json").
-            body("{}").
-            post("/api/allele/find?limit=10&page=0").
-            then().
-            statusCode(200).
-            body("totalResults", is(8));
-    }
-
-    @Test
     @Order(13)
-    public void alleleBulkUploadNoSymbolText() throws Exception {
-        String content = Files.readString(Path.of("src/test/resources/bulk/fms/02_allele/13_no_symbol_text.json"));
+    public void agmBulkUploadNoSubtype() throws Exception {
+        String content = Files.readString(Path.of("src/test/resources/bulk/fms/03_affected_genomic_model/13_no_subtype.json"));
         
         // upload file
         RestAssured.given().
             contentType("application/json").
             body(content).
             when().
-            post("/api/allele/bulk/allelefile").
+            post("/api/agm/bulk/agmfile").
             then().
             statusCode(200);
         
@@ -359,7 +360,7 @@ public class AlleleBulkUploadFmsITCase {
             when().
             header("Content-Type", "application/json").
             body("{}").
-            post("/api/allele/find?limit=10&page=0").
+            post("/api/agm/find?limit=10&page=0").
             then().
             statusCode(200).
             body("totalResults", is(8));
@@ -367,15 +368,15 @@ public class AlleleBulkUploadFmsITCase {
 
     @Test
     @Order(14)
-    public void alleleBulkUploadNoTaxonId() throws Exception {
-        String content = Files.readString(Path.of("src/test/resources/bulk/fms/02_allele/14_no_taxon_id.json"));
+    public void agmBulkUploadNoSynonyms() throws Exception {
+        String content = Files.readString(Path.of("src/test/resources/bulk/fms/03_affected_genomic_model/14_no_synonyms.json"));
         
         // upload file
         RestAssured.given().
             contentType("application/json").
             body(content).
             when().
-            post("/api/allele/bulk/allelefile").
+            post("/api/agm/bulk/agmfile").
             then().
             statusCode(200);
         
@@ -384,23 +385,23 @@ public class AlleleBulkUploadFmsITCase {
             when().
             header("Content-Type", "application/json").
             body("{}").
-            post("/api/allele/find?limit=10&page=0").
+            post("/api/agm/find?limit=10&page=0").
             then().
             statusCode(200).
-            body("totalResults", is(8));
+            body("totalResults", is(9));
     }
 
     @Test
     @Order(15)
-    public void alleleBulkUploadNoSynonyms() throws Exception {
-        String content = Files.readString(Path.of("src/test/resources/bulk/fms/02_allele/15_no_synonyms.json"));
+    public void agmBulkUploadNoTaxonId() throws Exception {
+        String content = Files.readString(Path.of("src/test/resources/bulk/fms/03_affected_genomic_model/15_no_taxon_id.json"));
         
         // upload file
         RestAssured.given().
             contentType("application/json").
             body(content).
             when().
-            post("/api/allele/bulk/allelefile").
+            post("/api/agm/bulk/agmfile").
             then().
             statusCode(200);
         
@@ -409,7 +410,7 @@ public class AlleleBulkUploadFmsITCase {
             when().
             header("Content-Type", "application/json").
             body("{}").
-            post("/api/allele/find?limit=10&page=0").
+            post("/api/agm/find?limit=10&page=0").
             then().
             statusCode(200).
             body("totalResults", is(9)); // taxonId is a required field so entity skipped in load;
@@ -417,15 +418,15 @@ public class AlleleBulkUploadFmsITCase {
 
     @Test
     @Order(16)
-    public void alleleBulkUploadNoSecondaryIds() throws Exception {
-        String content = Files.readString(Path.of("src/test/resources/bulk/fms/02_allele/16_no_secondary_ids.json"));
+    public void agmBulkUploadInvalidComponentAllele() throws Exception {
+        String content = Files.readString(Path.of("src/test/resources/bulk/fms/03_affected_genomic_model/16_invalid_agm_components_allele_id.json"));
         
         // upload file
         RestAssured.given().
             contentType("application/json").
             body(content).
             when().
-            post("/api/allele/bulk/allelefile").
+            post("/api/agm/bulk/agmfile").
             then().
             statusCode(200);
         
@@ -434,23 +435,23 @@ public class AlleleBulkUploadFmsITCase {
             when().
             header("Content-Type", "application/json").
             body("{}").
-            post("/api/allele/find?limit=10&page=0").
+            post("/api/agm/find?limit=10&page=0").
             then().
             statusCode(200).
-            body("totalResults", is(10));
+            body("totalResults", is(9));
     }
 
     @Test
     @Order(17)
-    public void alleleBulkUploadInvalidAlleleObjectRelationGene() throws Exception {
-        String content = Files.readString(Path.of("src/test/resources/bulk/fms/02_allele/17_invalid_allele_object_relation_gene.json"));
+    public void agmBulkUploadInvalidComponentZygosity() throws Exception {
+        String content = Files.readString(Path.of("src/test/resources/bulk/fms/03_affected_genomic_model/17_invalid_agm_components_allele_zygosity.json"));
         
         // upload file
         RestAssured.given().
             contentType("application/json").
             body(content).
             when().
-            post("/api/allele/bulk/allelefile").
+            post("/api/agm/bulk/agmfile").
             then().
             statusCode(200);
         
@@ -459,75 +460,75 @@ public class AlleleBulkUploadFmsITCase {
             when().
             header("Content-Type", "application/json").
             body("{}").
-            post("/api/allele/find?limit=10&page=0").
+            post("/api/agm/find?limit=10&page=0").
+            then().
+            statusCode(200).
+            body("totalResults", is(9));
+    }
+
+    // TODO: adjust count (and subsequent test counts) once loading and validation of STRs in place
+    @Test
+    @Order(18)
+    public void agmBulkUploadInvalidSequenceTargetingReagentId() throws Exception {
+        String content = Files.readString(Path.of("src/test/resources/bulk/fms/03_affected_genomic_model/18_invalid_str_id.json"));
+        
+        // upload file
+        RestAssured.given().
+            contentType("application/json").
+            body(content).
+            when().
+            post("/api/agm/bulk/agmfile").
+            then().
+            statusCode(200);
+        
+        // check entity count
+        RestAssured.given().
+            when().
+            header("Content-Type", "application/json").
+            body("{}").
+            post("/api/agm/find?limit=10&page=0").
             then().
             statusCode(200).
             body("totalResults", is(10));
     }
 
-    // TODO: adjust count (and subsequent test counts) once loading and validation of components in place
-    @Test
-    @Order(18)
-    public void alleleBulkUploadInvalidAlleleObjectRelationComponent() throws Exception {
-        String content = Files.readString(Path.of("src/test/resources/bulk/fms/02_allele/18_invalid_allele_object_relation_construct.json"));
-        
-        // upload file
-        RestAssured.given().
-            contentType("application/json").
-            body(content).
-            when().
-            post("/api/allele/bulk/allelefile").
-            then().
-            statusCode(200);
-        
-        // check entity count
-        RestAssured.given().
-            when().
-            header("Content-Type", "application/json").
-            body("{}").
-            post("/api/allele/find?limit=10&page=0").
-            then().
-            statusCode(200).
-            body("totalResults", is(11));
-    }
-
     @Test
     @Order(19)
-    public void alleleBulkUploadInvalidAlleleObjectRelationAssociationType() throws Exception {
-        String content = Files.readString(Path.of("src/test/resources/bulk/fms/02_allele/19_invalid_allele_object_relation_association_type.json"));
+    public void agmBulkUploadInvalidSubtype() throws Exception {
+        String content = Files.readString(Path.of("src/test/resources/bulk/fms/03_affected_genomic_model/19_invalid_subtype.json"));
         
         // upload file
         RestAssured.given().
             contentType("application/json").
             body(content).
             when().
-            post("/api/allele/bulk/allelefile").
+            post("/api/agm/bulk/agmfile").
             then().
-            statusCode(200);
+            statusCode(400);
         
         // check entity count
         RestAssured.given().
             when().
             header("Content-Type", "application/json").
             body("{}").
-            post("/api/allele/find?limit=10&page=0").
+            post("/api/agm/find?limit=10&page=0").
             then().
             statusCode(200).
-            body("totalResults", is(11));
+            body("totalResults", is(10));
     }
     
     // NOTE: validation currently only based on regex, not DB lookup
     @Test
     @Order(20)
-    public void alleleBulkUploadInvalidTaxonId() throws Exception {
-        String content = Files.readString(Path.of("src/test/resources/bulk/fms/02_allele/20_invalid_taxon_id.json"));
+    public void agmBulkUploadInvalidTaxonId() throws Exception {
+        String content = Files.readString(Path.of("src/test/resources/bulk/fms/03_affected_genomic_model/20_invalid_taxon_id.json"));
         
         // upload file
         RestAssured.given().
             contentType("application/json").
             body(content).
             when().
-            post("/api/allele/bulk/allelefile").
+            post("/api/agm/bulk/agmfile").
             then().
             statusCode(200);
         
@@ -536,10 +537,9 @@ public class AlleleBulkUploadFmsITCase {
             when().
             header("Content-Type", "application/json").
             body("{}").
-            post("/api/allele/find?limit=10&page=0").
+            post("/api/agm/find?limit=10&page=0").
             then().
             statusCode(200).
-            body("totalResults", is(11));
-    }
-    
+            body("totalResults", is(10));
+    }   
 }
